+++ src/KeyVault/KeyVault/Models/Key/PSDeletedKeyVaultKey.cs
@@ -43,7 +43,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
 
             Key = deletedKeyBundle.Key;
 
-            KeySize = JwkHelper.ConvertToRSAKey(Key)?.KeySize;
+            KeySize = JwkHelper.ComputeKeySize(Key);
 
             Attributes = new PSKeyVaultKeyAttributes(
                 deletedKeyBundle.Attributes.Enabled,
@@ -82,7 +82,11 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
             if(null != deletedKey.Key)
             {
                 Key = deletedKey.Key.ToTrack1JsonWebKey();
-                KeySize = JwkHelper.ConvertToRSAKey(Key)?.KeySize;
+                KeySize = deletedKey.Properties.KeySize ?? JwkHelper.ComputeKeySize(Key);
+            }
+            else
+            {
+                KeySize = deletedKey.Properties.KeySize;
             }
             Attributes = new PSKeyVaultKeyAttributes()
             {
@@ -122,7 +126,5 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
         {
             get { return Key?.CurveName; }
         }
-
-        public int? KeySize;
     }
 }